[Silver IV] Title: GCD 합, Time: 28 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/9613. GCD 합/GCD 합.cc
@@ -9,7 +9,7 @@ int main()
     cin >> t;
     while(t--){
       cin >> n;
-      vector<long long> arr(n);
+      vector<int> arr(n);
       long long sum = 0;
       for(int i = 0; i < n; i++){
         cin >> arr[i];
